infraestructure/requirements.txt
attrs==19.3.0
aws-cdk.aws-cloudwatch==1.44.0
aws-cdk.aws-ec2==1.44.0
aws-cdk.aws-events==1.44.0
aws-cdk.aws-iam==1.44.0
aws-cdk.aws-kms==1.44.0
aws-cdk.aws-logs==1.44.0
aws-cdk.aws-s3==1.44.0
aws-cdk.aws-ssm==1.44.0
aws-cdk.cdk-assets-schema==1.44.0
aws-cdk.cloud-assembly-schema==1.44.0
aws-cdk.core==1.44.0
aws-cdk.cx-api==1.44.0
aws-cdk.region-info==1.44.0
cattrs==1.0.0
constructs==3.0.3
-e git+git@github.com:BRCentralSA/amazon-eks-demo.git@699b55aa4c84b99a40c9f860b30a9ea7845f56c2#egg=infraestructure&subdirectory=infraestructure
jsii==1.6.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.2
